Move utility functions from api.py to utils.py
